refactor: utils 함수 파일로 분리
