+++ packages/assignment-7/src/App.tsx
@@ -34,6 +34,7 @@ import {
   VStack,
 } from '@chakra-ui/react';
 import { BellIcon, ChevronLeftIcon, ChevronRightIcon, DeleteIcon, EditIcon } from '@chakra-ui/icons';
+import { getWeekDates, formatWeek, getDaysInMonth, formatMonth } from './utils/date';
 
 type RepeatType = 'none' | 'daily' | 'weekly' | 'monthly' | 'yearly';
 
@@ -92,40 +93,6 @@ const fetchHolidays = (year: number, month: number) => {
     '2024-10-09': '한글날',
     '2024-12-25': '크리스마스',
   };
-};
-
-export const getDaysInMonth = (year: number, month: number) => {
-  return new Date(year, month + 1, 0).getDate();
-};
-
-export const getWeekDates = (date: Date) => {
-  const day = date.getDay();
-  const diff = date.getDate() - day + (day === 0 ? -6 : 1);
-  const monday = new Date(date.setDate(diff));
-  const weekDates = [];
-  for (let i = 0; i < 7; i++) {
-    const nextDate = new Date(monday);
-    nextDate.setDate(monday.getDate() + i);
-    weekDates.push(nextDate);
-  }
-  return weekDates;
-};
-
-export const formatWeek = (date: Date): string => {
-  const year = date.getFullYear();
-  const month = date.getMonth() + 1;
-  const weekNumber = Math.ceil(date.getDate() / 7);
-  return `${year}년 ${month}월 ${weekNumber}주`;
-};
-
-export const formatMonth = (date: Date): string => {
-  const year = date.getFullYear();
-  const month = date.getMonth() + 1;
-  return `${year}년 ${month}월`;
-};
-
-export const isDateInRange = (date: Date, startDate: Date, endDate: Date): boolean => {
-  return date >= startDate && date <= endDate;
 };
 
 function App() {
